+++ test/http/features/instrumentation_test.rb
@@ -72,10 +72,40 @@ describe HTTP::Features::Instrumentation do
       null_instrumenter.finish("test", {})
     end
 
+    it "#instrument is callable without a block" do
+      null_instrumenter.instrument("test")
+    end
+
     it "#instrument yields the block when given" do
       result = null_instrumenter.instrument("test") { :yielded }
 
       assert_equal :yielded, result
+    end
+  end
+
+  context "with an instrumenter that unconditionally yields" do
+    let(:yielding_instrumenter) do
+      Class.new do
+        def instrument(_name, _payload = {})
+          yield
+        end
+
+        def start(_name, _payload = {}); end
+        def finish(_name, _payload = {}); end
+      end.new
+    end
+
+    let(:feature) { HTTP::Features::Instrumentation.new(instrumenter: yielding_instrumenter) }
+    let(:request) do
+      HTTP::Request.new(verb: :get, uri: "https://example.com/", headers: {})
+    end
+
+    it "does not raise LocalJumpError in wrap_request" do
+      feature.wrap_request(request)
+    end
+
+    it "does not raise LocalJumpError in on_error" do
+      feature.on_error(request, HTTP::TimeoutError.new)
     end
   end